fix: mergeField 키 버그 수정 & refactor: 히스토리 기록 통일
 - ChatHistoryStore에 recordChatTurn/recordResolvedTurn 추가로 기록 조립 책임 이동
 - ChatService/ChatResolveService에서 히스토리 포맷 로직 제거
 - mergeField가 Kotlin 프로퍼티 이름을 써서 missing_fields가 비워지지 않던 버그를 JSON 키(snake_case)로 교정. 완료 action에 남던 message도 함께 제거.
